mm/address_space: Increase stack size for attestation report
The current attestation report implementation requires a bigger stack
size. Without increasing the stack size the guest panics with a
Double-Fault.

[SVSM] Validating 0x000000000080a000-0x0000000000820000
[SVSM] Flash region 0 at 0x00000000ffc84000 size 00000000000037c000
[SVSM] Flash region 1 at 0x00000000ffaf5000 size 00000000000010b000
[SVSM] ERROR: Panic: CPU[0] panicked at 'Double-Fault at RIP 0xffffff800008feec RSP: 0xffffff00001ffff8 CR2: 0xffffff00001fffe8', src/cpu/idt.rs:169:9
[SVSM] ---BACKTRACE---:
[SVSM]   [0xffffff800002cfed]
[SVSM]   [0xffffff800000003f]
[SVSM]   Invalid frame
[SVSM] ---END---

The problem goes away if the stack size is at least 10 pages. We are
setting it to 12 pages to have some extra space.

This does not change the stack size for FEATURES=enable-gdb. The guest
seems to be working fine with the current 16 pages allocated for the
stack.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>