psp: Add SnpGuestRequestMsg structure and implementation
The AMD secure processor (PSP) provides SNP Guest Request services, e.g.
the attestation report service. To request them we have to wrap the
respective service structure into a SnpGuestRequestMsg structure and
send it encrypted to the PSP. The response will also be encrypted and
wrapped in a SnpGuestRequestMsg structure.

This uses the RustCrypto AES_GCM to encrypt and decrypt the
SnpGuestRequestMsg payload.

This also adds the SnpGuestRequestExtData. When we request an extended
attestation report, the PSP will provide the attestation report and the
hypervisor will use the SnpGuestRequestExtData.buffer to store the
certificate chain used to verify the attestation report.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>